Change hard drive type to VMDK - Dynamic and split
I would recommend using VMDK instead of VDI as it's compatible with other virtualising apps like VMware Player for example. I would also split in 2GB files so it's easier to move around, copy, upload on cloud etc and also instead of fixed I would create the drive as dynamic in case someone creates a large drive (80GB+ for example), that's not pre-allocated.